requirements.txt
python-telegram-bot==13.15
requests
python-dotenv
yfinance
pandas
numpy
pandas_ta
scikit-learn
beautifulsoup4
apscheduler
matplotlib
prophet
